Extract MyToolsLanding from ToolBox into its own component
